Refactor propagation further
Structure the propagation routines in a more modular fashion to allow
for better reuse. Each different stage of the analysis is now cleanly
separated from the rest. Function and record names should be more
intuitive as well. The documentation was also revised and expanded.

There is a small performance improvement too, gained by keeping a unique
id instead of each cycle in the cycle map, and by caching the map in the
state record instead of recomputing it at each iteration.