Fix bug in name generation of nested functions
The bug was caused by using only the name of parent functions, which is not
unique in Erlang modules. This caused name clashes, overwriting analysis
results of other nested functions.

The simplest solution is to use the parent's arity as well, when constructing
the unique name.

More functions show up in analysis results now, and some functions will also
have different results.

Besides the bug, refactor name generation so that new atoms are only introduced
when necessary.