Fix recursion bug, handle nested funs better
Recursive calls are important when it comes to higher order functions,
since they may influence their purity decisively. It was incorrect to
ignore them all this time.

The new approach is to keep track of all recursive calls in the
dependency list and then remove the ones not relevant to the analysis
before propagation. For non-higher-order functions just remove all of
them. For HOFs however we can only remove those which include the same
set of arguments the function depends on (we only track HOFs which
depend on their arguments anyway, the rest are unresolvable). Call site
analysis is also modified to handle recursive calls with concrete higher
order arguments.

This was a serious bug since it allowed for false positives. Its impact
however was very small, since such functions are very rare (only one
such example was detected in OTP, `dets_utils:leafs_to_nodes/4').

Dependencies on free variables in nested functions are now propagated
much better, allowing for arbitrary levels of nesting and supporting
recursive calls as well. The approach taken flattens dependencies to
nested functions while transforming them as appropriate, for instance
mapping free variables to function arguments in the parent. The analysis
is somewhat faster now too, presumably because of this flattening.

Add examples as test cases and reorganise the tests for HOFs.