Minor refactoring and unit tests
Refactor the argument matching function and fix a minor bug.

Add some unit tests and a simple script and Makefile rule to run them.
The setup is not ideal (code needs to be recompiled with -DTEST) but it
will work for the time being. The tests are isolated in a separate file
to not clutter the source code. By including this file instead of making
use of the built-in mechanism by eunit, non-exported functions can be
tested as well.