Add simple termination analysis
Evaluation in bounded time is one of the prerequisites for guard
expressions. To this end, a crude and very conservative termination
analysis is added, to be used as a placeholder if necessary.

A function is considered non-terminating if it calls itself, has a
potentially infinite receive statement, or calls another
non-terminating function. All built-in functions are considered
terminating.

A command line option was added for running the termination analysis.
API-wise it simply provides a different propagation step. Both analyses
take advantage of the same traversal data and may reside in the same
PLT.