Refactor and enhance command line parsing slightly
Add a post-processing step in command line parsing so that duplicate
options and overrides can be handled. This has an effect on PLT caching,
since some options are used as keys. The previous version cached
termination analysis results under `[termination,{purelevel,_}]' when it
should have been simply `[termination]', which is the current behaviour.

Switch to *_tests.hrl from *_tests.erl for unit tests so that edoc will
not complain. Move the `-ifdef(TEST)' macros to the files calling
`-include', so that they may compile without the test file present.

Fix the Makefile rule to provide all modules to the unit test script
correctly, not just the first one.

Fix minor typos and update documentation.