Reduce memory usage in propagation
The memory usage of the propagation stage of the analysis was pretty
high, requiring huge amounts of memory when running both analyses on OTP
with --level 3. The main cause are the reason strings kept for impure
functions.

The issue is addressed by not keeping track of reasons by default.
Instead the user has to provide an appropriate option. Even then, the
representation is changed from strings (lists of integers) to binaries,
bringing down the memory usage to something more bearable. One approach
that is likely to be even better, would be to create the message on the
fly only when it's necessary. This would miss call site dependencies,
but these could be handled as they are now, they are few anyway.

Besides the reduced memory footprint, there is also a performance
increase.