Enhance termination analysis
Detect functions which recursively consume one of their arguments. These
are guaranteed to terminate (or crash). Currently this works for some
cases of list and binary processing functions, for instance many of the
functions in the lists stdlib module.

The approach taken is to track variables which are a strict subset of
any of the function's arguments at each case clause, and add that
information for recursive calls. If all the recursive dependencies of a
function have at least one such subset in common (at the same argument
position), they can be removed from the dependency list prior to
propagation.

The percentage of terminating functions increases from ~18% to ~25% for
OTP R14B.