Faster combined analysis
Merge the two propagation functions into one, providing for a
significant speedup when both analyses are requested. Add
`combined_test` script to ensure the results are consistent with running
the two analyses separately like before (has to be run manually with any
beam files as arguments).

Fix bug in purity_plt, cached results of combined analysis where saved
as results of purity analysis.

Fix minor bug in termination analysis regarding the removal of recursion
for calls with subsets. The case of higher order functions calling
themselves with an unknown function was not handled correctly. This
could potentially creates false positives in very rare cases. Update
tests to catch it. Take the opportunity to refactor the recursion
handling code.