Fix bug in promotion of nested dependencies
The way nested dependencies are tracked it is possible to include the
function itself in the nested list (in case of letrecs), but there was
no check to exclude that case. This caused a crash in the
`couch_changes` module of CouchDB, where the function itself was being
mapped and a {sub,_} was included. There should be no impact on results,
since nothing would get promoted in such a case anyway.