Fix bug with module name conflicts
During parallel traversal the table was not cleared from any modules
with the same name, since the analysis starts from an empty table and
the results are merged. The merge functions was changed to remove any
previous entries prior to merging.

The bug was spotted while analysing Wings3D which ships with its own
array module.