Optimise table merging, add test script
The fix caused a minor performance drop during the traversal for big
PLTs because the initial table was traversed each time during the merge.
It is now back to similar levels.

Add script for making sure that analysing sequentially and in parallel
produce the same results, along with a very simple test case were two
modules with the same name are analysed.

Update the runtests script to include the new test, as well as the one
for the previous bug with combined analysis.