Update tests, fix HOF limitation
Update all purity-related test cases to handle the new results, and a
slightly simpler syntax for describing HOFs. It is now unnecessary to
have separate results for different purity levels. However, certain
subtle test cases which depend on the reason for which a function was
characterised impure are now indistinguishable.

Termination-related test cases are left as-is for the moment, and
subsequently cause 2 failing tests.

Do not restore self-dependencies after the contamination pass, only
{arg,_} tuples. This change does not really affect HOFs, since they are
unresolvable anyway, but it enables resolution of functions which
acquired self-dependencies because of HOF unfolding, roughly 0.1% of
OTP.

By moving self-dependency removal to the pre-processing stage, it is now
possible to resolve some uncommon corner cases as described in
test/nested. However, we now miss a case previously resolved, where
indirect HOFs and HOFs being passed concrete arguments are combined
(only one such case actually resolved in the whole of OTP, several
others were still unresolvable before).