Optimise contamination loop
Achieve significant speedup in contamination (8x for termination
analysis and ~1.5x for purity analysis) by keeping track of the working
set as a nested list with possible duplicates, instead of an ordered
set.

Since the working set is not used in the inner loop, only incremented,
it is possible to move the working set invariant (list of unique
functions) to the outer loop, where it is actually used.

Performance is now similar (if not better) to the old algorithm.