Update broken test
The termination result of m/2 in test/higher was incorrectly left to
`s', instead of the more accurate `>= p [1]'. While there is no
difference in practice, the significance lies in telling apart recursion
(which is resolved correctly here) from unhandled HOFs.

The issue was uncovered because the recursion removal functions of the
previous algorithm did not work on unhandled HOFs.