Update plt and utils
For the moment rewrite into separate modules for easier testing.

Besides updating any relevant functions to work on the new values of
purity, there have been some minor enhancements and fixes.

For instance, the PLT update function did not check that the caches it
kept were consistent. Also, the dependency resolution of affected
modules was not correct, as it only took into account direct
dependencies.