Fix bug in termination analysis
Setting all BIFs as terminating is wrong. Part of this is fixed now,
namely honoring dependency lists and unresolvable BIFs. Still there is
work to be done here, since the current approach cannot express other
types of non-terminating BIFs at the moment.

This is not very serious in the sense that any such BIFs are most likely
impure functions too (e.g. port calls), so there should be no impact on
the combined results of the analysis (which is the ultimate goal), just
the termination part.

Note also that two commits ago a significant fix went unnoticed. Some
BIFs were not properly converted to the new values for termination
(those marked with `false' in the cheat sheets), and were erroneously
marked as non-terminating.