Analyse actual dependencies, not the whole PLT
Optimise analysis in the presence of a PLT. Only analyse the necessary
dependencies, and not the whole PLT as past versions did. This cannot
always be avoided however, for instance when there is no result table
present for the specific set of options. Another problem is when the PLT
is to be updated. In this case, extract only those functions which are
still unresolved in the result table.

Take the opportunity to decouple the first stage of the analysis from
the PLT and any analysis options, since neither is actually necessary.
Move these dependencies to the propagation stage instead. Change to more
descriptive function names in the collect module.

Add functions to export a PLT into a form more suitable for debugging,
as well as a plain text file.

Refactor the timing functions and move them into the utils module,
allowing them to be used by external scripts, etc.

Add a PLT test script to test the new analysis and PLT update scheme, as
well as cover most of the new PLT code.

The analysis runs much faster now when analysing a small set of modules
against a large PLT. The remaining bottlenecks are the PLT loading and
updating functions, which have a much smaller impact and can also be
aided by caching at the operating system level (the loading part).